Move Luanti rotation matrix helpers to own header
